+++ test/mocks/api/mocks.cc
@@ -27,7 +27,7 @@ Event::DispatcherPtr MockApi::allocateDispatcher(Buffer::WatermarkFactoryPtr&& w
 }
 
 MockOsSysCalls::MockOsSysCalls() {
-  ON_CALL(*this, close(_)).WillByDefault(Invoke([](int fd) {
+  ON_CALL(*this, close(_)).WillByDefault(Invoke([](int64_t fd) {
     const int rc = ::close(fd);
     return SysCallIntResult{rc, errno};
   }));
@@ -35,8 +35,8 @@ MockOsSysCalls::MockOsSysCalls() {
 
 MockOsSysCalls::~MockOsSysCalls() = default;
 
-SysCallIntResult MockOsSysCalls::setsockopt(int sockfd, int level, int optname, const void* optval,
-                                            socklen_t optlen) {
+SysCallIntResult MockOsSysCalls::setsockopt(int64_t sockfd, int level, int optname,
+                                            const void* optval, socklen_t optlen) {
   ASSERT(optlen == sizeof(int));
 
   // Allow mocking system call failure.
@@ -48,7 +48,7 @@ SysCallIntResult MockOsSysCalls::setsockopt(int sockfd, int level, int optname, 
   return SysCallIntResult{0, 0};
 };
 
-SysCallIntResult MockOsSysCalls::getsockopt(int sockfd, int level, int optname, void* optval,
+SysCallIntResult MockOsSysCalls::getsockopt(int64_t sockfd, int level, int optname, void* optval,
                                             socklen_t* optlen) {
   ASSERT(*optlen == sizeof(int));
   int val = 0;
